refactor: change core layer to be totally isolated and many more
- refreshed packages
- always use 'production' environment for tests
- core layer returns only dtos
- core layer has dedicated exceptions: contract and missing data issues
- dedicated class to choose random movement
- repository implementation is in infra layer
- all core exception are captured in api layer and then treated

+++ src/Core/Models/PlayerBoard.cs
@@ -2,9 +2,9 @@ namespace TicTacToeCSharpPlayground.Core.Models
 {
     public class PlayerBoard : StandardEntity
     {
-        public long PlayerId { get; set; }
+        public int PlayerId { get; set; }
         public Player Player { get; set; }
-        public long BoardId { get; set; }
+        public int BoardId { get; set; }
         public Board Board { get; set; }
     }
 }
