+++ src/Core/Models/StandardEntity.cs
@@ -5,7 +5,7 @@ namespace TicTacToeCSharpPlayground.Core.Models
 {
     public abstract class StandardEntity
     {
-        [Key] public long Id { get; set; }
+        [Key] public int Id { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
